Update DBALibrary to check logging table and SMTP server
The script now checks if the logging table is accessible.  If it is not logging to the database is disabled so the script should still now run without error when the logging table is not yet in place.

It also checks that the SMTP server is listening.  This is not a full check for functionality of emails as this is not possible to test without sending a test email.  e.g. the SMTP server could have a whitelist.